refactor : Path에 대한 에러 핸들링 로직 패키지 분리, 안 쓰는 코드 삭제
